+++ bindings/angular1/directives/actionSheet.js
@@ -1,0 +1,141 @@
+/**
+ * @element ons-action-sheet
+ */
+
+/**
+ * @attribute var
+ * @initonly
+ * @type {String}
+ * @description
+ *  [en]Variable name to refer this action sheet.[/en]
+ *  [ja]このアクションシートを参照するための名前を指定します。[/ja]
+ */
+
+/**
+ * @attribute ons-preshow
+ * @initonly
+ * @type {Expression}
+ * @description
+ *  [en]Allows you to specify custom behavior when the "preshow" event is fired.[/en]
+ *  [ja]"preshow"イベントが発火された時の挙動を独自に指定できます。[/ja]
+ */
+
+/**
+ * @attribute ons-prehide
+ * @initonly
+ * @type {Expression}
+ * @description
+ *  [en]Allows you to specify custom behavior when the "prehide" event is fired.[/en]
+ *  [ja]"prehide"イベントが発火された時の挙動を独自に指定できます。[/ja]
+ */
+
+/**
+ * @attribute ons-postshow
+ * @initonly
+ * @type {Expression}
+ * @description
+ *  [en]Allows you to specify custom behavior when the "postshow" event is fired.[/en]
+ *  [ja]"postshow"イベントが発火された時の挙動を独自に指定できます。[/ja]
+ */
+
+/**
+ * @attribute ons-posthide
+ * @initonly
+ * @type {Expression}
+ * @description
+ *  [en]Allows you to specify custom behavior when the "posthide" event is fired.[/en]
+ *  [ja]"posthide"イベントが発火された時の挙動を独自に指定できます。[/ja]
+ */
+
+/**
+ * @attribute ons-destroy
+ * @initonly
+ * @type {Expression}
+ * @description
+ *  [en]Allows you to specify custom behavior when the "destroy" event is fired.[/en]
+ *  [ja]"destroy"イベントが発火された時の挙動を独自に指定できます。[/ja]
+ */
+
+/**
+ * @method on
+ * @signature on(eventName, listener)
+ * @description
+ *   [en]Add an event listener.[/en]
+ *   [ja]イベントリスナーを追加します。[/ja]
+ * @param {String} eventName
+ *   [en]Name of the event.[/en]
+ *   [ja]イベント名を指定します。[/ja]
+ * @param {Function} listener
+ *   [en]Function to execute when the event is triggered.[/en]
+ *   [ja]イベントが発火された際に呼び出されるコールバックを指定します。[/ja]
+ */
+
+/**
+ * @method once
+ * @signature once(eventName, listener)
+ * @description
+ *  [en]Add an event listener that's only triggered once.[/en]
+ *  [ja]一度だけ呼び出されるイベントリスナーを追加します。[/ja]
+ * @param {String} eventName
+ *   [en]Name of the event.[/en]
+ *   [ja]イベント名を指定します。[/ja]
+ * @param {Function} listener
+ *   [en]Function to execute when the event is triggered.[/en]
+ *   [ja]イベントが発火した際に呼び出されるコールバックを指定します。[/ja]
+ */
+
+/**
+ * @method off
+ * @signature off(eventName, [listener])
+ * @description
+ *  [en]Remove an event listener. If the listener is not specified all listeners for the event type will be removed.[/en]
+ *  [ja]イベントリスナーを削除します。もしlistenerパラメータが指定されなかった場合、そのイベントのリスナーが全て削除されます。[/ja]
+ * @param {String} eventName
+ *   [en]Name of the event.[/en]
+ *   [ja]イベント名を指定します。[/ja]
+ * @param {Function} listener
+ *   [en]Function to execute when the event is triggered.[/en]
+ *   [ja]削除するイベントリスナーの関数オブジェクトを渡します。[/ja]
+ */
+
+(function() {
+  'use strict';
+
+  /**
+   * Action sheet directive.
+   */
+  angular.module('onsen').directive('onsActionSheet', function($onsen, ActionSheetView) {
+    return {
+      restrict: 'E',
+      replace: false,
+      scope: true,
+      transclude: false,
+
+      compile: function(element, attrs) {
+
+        return {
+          pre: function(scope, element, attrs) {
+            var actionSheet = new ActionSheetView(scope, element, attrs);
+
+            $onsen.declareVarAttribute(attrs, actionSheet);
+            $onsen.registerEventHandlers(actionSheet, 'preshow prehide postshow posthide destroy');
+            $onsen.addModifierMethodsForCustomElements(actionSheet, element);
+
+            element.data('ons-action-sheet', actionSheet);
+
+            scope.$on('$destroy', function() {
+              actionSheet._events = undefined;
+              $onsen.removeModifierMethods(actionSheet);
+              element.data('ons-action-sheet', undefined);
+              element = null;
+            });
+          },
+          post: function(scope, element) {
+            $onsen.fireComponentEvent(element[0], 'init');
+          }
+        };
+      }
+    };
+  });
+
+})();