+++ bindings/react/demo/index.js
@@ -19,6 +19,7 @@ import ListExample from './examples/List';
 import LazyListExample from './examples/LazyList';
 import TabbarExample from './examples/Tabbar';
 import AlertDialogExample from './examples/AlertDialog';
+import ActionSheetExample from './examples/ActionSheet';
 import ToastExample from './examples/Toast';
 import SplitterExample from './examples/Splitter';
 import InputExample from './examples/Input';
@@ -136,6 +137,10 @@ class Examples extends React.Component {
       {
         title: 'Alert dialog',
         component: AlertDialogExample
+      },
+      {
+        title: 'Action sheet',
+        component: ActionSheetExample
       },
       {
         title: 'Toast',