+++ bindings/react/src/index.js
@@ -1,3 +1,5 @@
+import ActionSheet from './components/ActionSheet.jsx';
+import ActionSheetButton from './components/ActionSheetButton.jsx';
 import AlertDialog from './components/AlertDialog.jsx';
 import BackButton from './components/BackButton.jsx';
 import BottomToolbar from './components/BottomToolbar.jsx';
@@ -42,6 +44,8 @@ import ToolbarButton from './components/ToolbarButton.jsx';
 import RouterUtil from './RouterUtil';
 
 export {
+  ActionSheet,
+  ActionSheetButton,
   AlertDialog,
   BackButton,
   BottomToolbar,