+++ bindings/vue/src/index.js
@@ -53,8 +53,9 @@ const install = (Vue, params = {}) => {
    */
   Vue.prototype.$ons = Object.keys(ons)
     .filter(k => [
-      /^enable/,
       /^disable/,
+      /^enable/,
+      /^open/,
       /^set/,
       /animit/,
       /Element$/,