+++ core/src/setup.js
@@ -4,6 +4,8 @@ import ons from './ons/ons';
 
 import TemplateElement from './elements/ons-template';
 import IfElement from './elements/ons-if';
+import ActionSheetElement from './elements/ons-action-sheet';
+import ActionSheetButtonElement from './elements/ons-action-sheet-button';
 import AlertDialogElement from './elements/ons-alert-dialog';
 import BackButtonElement from './elements/ons-back-button';
 import BottomToolbarElement from './elements/ons-bottom-toolbar';
@@ -47,6 +49,8 @@ import RangeElement from './elements/ons-range';
 
 ons.TemplateElement = TemplateElement;
 ons.IfElement = IfElement;
+ons.ActionSheetElement = ActionSheetElement;
+ons.ActionSheetButtonElement = ActionSheetButtonElement;
 ons.AlertDialogElement = AlertDialogElement;
 ons.BackButtonElement = BackButtonElement;
 ons.BottomToolbarElement = BottomToolbarElement;
@@ -97,7 +101,11 @@ window.addEventListener('load', () => {
 window.addEventListener('DOMContentLoaded', () => {
   ons._deviceBackButtonDispatcher.enable();
   ons._defaultDeviceBackButtonHandler = ons._deviceBackButtonDispatcher.createHandler(window.document.body, () => {
-    navigator.app.exitApp();
+    if (Object.hasOwnProperty.call(navigator, 'app')) {
+      navigator.app.exitApp();
+    } else {
+      console.warn('Could not close the app. Is \'cordova.js\' included?\nError: \'window.navigator.app\' is undefined.');
+    }
   });
   document.body._gestureDetector = new ons.GestureDetector(document.body);
 }, false);
@@ -105,6 +113,15 @@ window.addEventListener('DOMContentLoaded', () => {
 // setup loading placeholder
 ons.ready(function() {
   ons._setupLoadingPlaceHolders();
+
+  // Simulate Device Back Button on ESC press
+  if (!ons.platform.isWebView()) {
+    document.body.addEventListener('keydown', function(event) {
+      if (event.keyCode === 27) {
+        ons._deviceBackButtonDispatcher.fireDeviceBackButtonEvent();
+      }
+    })
+  }
 });
 
 // viewport.js