+++ bindings/vue/CHANGELOG.md
@@ -1,11 +1,15 @@
 CHANGELOG
 ====
 
+dev
 2.0.0-beta.4
 ----
 
 ### New Features
 
+* VOnsActionSheet: New component.
+* VOnsActionSheetButton: New component.
+* $ons.actionSheet: Suppot shortcut method for VOnsActionSheet.
 * Vue 2.3.0: `sync` modifier is now supported for `index`, `visible` and `open` props. See breaking changes section.
 
 ### Bug Fixes